# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Boisolo Api",
        "description": "This is a sample API documentation",
        "contact": {
            "email": "dev@boisolo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sensors/{token}": {
            "get": {
                "tags": [
                    "Sensores"
                ],
                "summary": "Obtener datos del sensor por token",
                "operationId": "f5faa6b35d2bdbf9c74ca13977f9b46b",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token del sensor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del sensor obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Sensor 1"
                                        },
                                        "sensor_type": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "mqtt_topic_sensor": {
                                            "type": "string",
                                            "example": "sensor/topic/1"
                                        },
                                        "count_total": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Sensor not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/sensors/insert": {
            "post": {
                "tags": [
                    "Sensores"
                ],
                "summary": "Insertar datos del sensor",
                "description": "Inserta datos en el sensor, verificando si necesita invertir el valor según su configuración.",
                "operationId": "b588a760b55c4b906e3abed7ad38cca1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "value": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sensor": {
                                        "type": "string",
                                        "example": "sensor/topic"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos procesados correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Datos procesados correctamente"
                                        },
                                        "sensor": {
                                            "type": "string",
                                            "example": "Sensor 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Sensor no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}